backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# LLM & AI - LangChain 1.0+
# 核心包
langchain>=1.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# 模型提供商集成
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.3.0

# 原生 SDK (用于直接调用)
openai>=1.10.0
anthropic>=0.18.0

# 兼容旧版功能 (如需要使用 chains, retrievers 等)
# langchain-classic>=0.1.0

# RAG & Vector DB
chromadb>=0.4.22
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
pypdf>=4.0.1
python-docx>=1.1.0
markdown>=3.5.2
beautifulsoup4>=4.12.3

# Embeddings
tiktoken==0.5.2

# MCP
httpx==0.26.0
aiohttp==3.9.1

# Data & Storage
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
pytest-cov==4.1.0

# Additional dependencies for LangChain 1.0
dataclasses-json>=0.6.0
